type casts for VC++ 2005 Express compilation

+++ src/drivers/berniw/pathfinder.cpp
@@ -664,7 +664,7 @@ void Pathfinder::plan(MyCar* myc)
 #ifdef PATH_K1999
 	/* compute path */
 	for (int step = 128; (step /= 2) > 0;) {
-		for (int i = 100 * int(sqrt(step)); --i >= 0;) smooth(step);
+		for (int i = 100 * int(sqrt((double)step)); --i >= 0;) smooth(step);
 		interpolate(step);
 	}
 #endif	// PATH_K1999
